Just a proposal to use constant instead of define
This is just a proposal to use constant instead of define because const seems to be faster than define and more cleaner. This could be used also for database_tables.php, configure.php and language file definitions. From php 5.3 const can be used in the global scope and is case sensitive.

http://stackoverflow.com/questions/2447791/define-vs-const